GUI update
Stopped at the visible login failed label
